move shared function to a library
move selectOntologyNode from manage-explorer.js to utils/data/block-values.js
and use it also in ontologies.js
