+++ frontend/app/components/panel/manage-explorer.js
@@ -50,6 +50,7 @@ import {
   blockValuesTree,
   blockFeatureOntologiesTreeEmbeddedFn,
   idParentNodeMap,
+  selectOntologyNode,
 } from '../../utils/data/block-values';
 
 import ManageBase from './manage-base'
@@ -1760,28 +1761,7 @@ export default ManageBase.extend({
   },
   //----------------------------------------------------------------------------
 
-  /** user has clicked on a entry-expander in an ontology tree. */
-  selectOntologyNode(nodeText, values, event) {
-    dLog('selectOntologyNode', nodeText, values, event.target);
-    let ontologyId = values?.id || values?.name && ontologyIdFromIdText(values.name);
-    if (ontologyId) {
-      let colour = this.get('ontology').ontologyClick(ontologyId);
-      let target = event?.target;
-      if (target && colour) {
-        /** probably "" or undefined */
-        let previousColour = target.style.background;
-        target.style.background = colour;
-        /** For qtlColourHierarchy() it is useful to show the clicked element
-         * colour briefly then clear it - simpler than clearing colours when
-         * another node is clicked :
-         *   later(() => target.style.background = previousColour, 2 * 1000);
-         * Now that qtlColourLevel() is used, the background-color is provided
-         * by entry-expander : valuesColour(), which clears the background-color when
-         * needed.
-         */
-      }
-    }
-  },
+  selectOntologyNode,
 
   // ---------------------------------------------------------------------------
