+++ frontend/app/components/panel/view-controls.js
@@ -311,7 +311,7 @@ export default Component.extend({
   /** ditto, 
    * controls.view.pathControlActiveSample
    */
-  pathSample : expRangeInitial(400, expRangeBase(100, 10000)),
+  pathSample : expRangeInitial(4/*00*/, expRangeBase(100, 10000)),
 
   pathControlActiveDensity : computed('pathDensityActive', 'pathDensity', function () {
     let active = this.get('pathDensityActive'),
